+++ lib/amazing_print/core_ext/string.rb
@@ -40,4 +40,9 @@ class String
 
   alias black grayish
   alias pale  whiteish
+
+  # Remove ANSI color codes.
+  def uncolor
+    gsub(/\e\[[0-9;]*m/, '')
+  end
 end